codex-harness/harness/orchestrator/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
httpx>=0.27.0
pydantic>=2.9.0
azure-storage-blob>=12.19.0
azure-identity>=1.16.0
